changing calculate structure

+++ commission_calculator_backend/db/migrate/20191231053607_create_sales.rb
@@ -2,10 +2,10 @@ class CreateSales < ActiveRecord::Migration[5.2]
   def change
     create_table :sales do |t|
       t.integer :commission_rate
-      t.float :price
-      t.float :quantity
+      t.integer :price
+      t.integer :quantity
       t.integer :product_id
-      t.float :commission_amount
+      t.integer :commission_amount
       t.string :commission_type
 
       t.timestamps
